day4/google_adk/requirements.txt
google-adk[database]
yfinance
psutil
litellm
google-generativeai
python-dotenv
aiosqlite